refactor: combined onboarding and user features
