+++ lib/app.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_template_core/core/config/router.dart';
 import 'package:flutter_template_core/core/theme/theme.dart';
-import 'package:flutter_template_core/features/user/controllers/theme_controller.dart';
+import 'package:flutter_template_core/features/auth/controllers/theme_controller.dart';
 import 'package:flutter_template_core/flavors.dart';
 
 class MainApp extends ConsumerWidget {
