+++ lib/core/common/ui/screens/settings_screen.dart
@@ -5,9 +5,9 @@ import 'package:flutter_template_core/core/common/ui/components/flex_list_tile.d
 import 'package:flutter_template_core/core/common/ui/components/section.dart';
 import 'package:flutter_template_core/core/extensions/ui_extensions.dart';
 import 'package:flutter_template_core/core/theme/app_layout.dart';
-import 'package:flutter_template_core/features/user/ui/containers/theme_selector.dart';
-import 'package:flutter_template_core/features/user/ui/containers/user_info_card.dart';
-import 'package:flutter_template_core/features/user/ui/screens/profile_screen.dart';
+import 'package:flutter_template_core/features/auth/ui/containers/theme_selector.dart';
+import 'package:flutter_template_core/features/auth/ui/containers/user_info_card.dart';
+import 'package:flutter_template_core/features/auth/ui/screens/profile_screen.dart';
 import 'package:go_router/go_router.dart';
 
 class SettingsScreen extends ConsumerWidget {
