+++ lib/core/config/router.dart
@@ -2,11 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:flutter_swipe_action_cell/core/swipe_action_navigator_observer.dart';
 import 'package:flutter_template_core/core/common/ui/screens/error_screen.dart';
 import 'package:flutter_template_core/core/common/ui/screens/main_screen.dart';
-import 'package:flutter_template_core/features/onboarding/providers.dart';
-import 'package:flutter_template_core/features/onboarding/ui/screens/onboarding_screen.dart';
+import 'package:flutter_template_core/features/auth/providers.dart';
+import 'package:flutter_template_core/features/auth/ui/screens/onboarding_screen.dart';
 import 'package:flutter_template_core/features/posts/ui/screens/post_create_screen.dart';
 import 'package:flutter_template_core/features/posts/ui/screens/post_edit_screen.dart';
-import 'package:flutter_template_core/features/user/ui/screens/profile_screen.dart';
+import 'package:flutter_template_core/features/auth/ui/screens/profile_screen.dart';
 import 'package:go_router/go_router.dart';
 
 /// Main router for the Example app
