+++ lib/features/auth/ui/components/onboarding_getting_started.dart
