+++ lib/features/auth/ui/components/theme_display.dart
