--- lib/features/onboarding/controllers/user_form_controller.dart
@@ -1,17 +1,0 @@
-import 'package:flutter_template_core/features/onboarding/data/models/user_form_model.dart';
-import 'package:flutter_template_core/features/user/controllers/user_create_controller.dart';
-import 'package:riverpod_annotation/riverpod_annotation.dart';
-
-part 'user_form_controller.g.dart';
-
-@riverpod
-class UserFormController extends _$UserFormController {
-  @override
-  UserForm build() {
-    return UserForm(UserForm.formElements(const User()), null);
-  }
-
-  void create() {
-    ref.read(userCreateControllerProvider.notifier).handle(state);
-  }
-}
